+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 3.1.0
 
+- improved the overall documentation
+- fixed a bug where `ChangeNotifierProvider.value` didn't update dependents
+  when the `ChangeNotifier` instance changed.
 - `Consumer` can now be used inside `MultiProvider`
   ```dart
   MultiProvider(